Do not ignore system threads in pfs processlist
Summary:
We are trying to maintain parity with show processlist.
Pfs.processlist implementation can ignore system threads but right after it has the provision to consider the system threads too, so removing the initial check.

I am keeping the check on lack of processlist id since those are primarily daemons etc. which are also ignored by legacy show processlist.

Here's what the original processlist ignores:

      {
        MUTEX_LOCK(grd, &inspect_thd->LOCK_thd_protocol);
        if ((!inspect_thd->get_protocol()->connection_alive() &&
             !inspect_thd->system_thread) ||
            (user && (inspect_thd->system_thread || !inspect_sctx_user.str ||
                      strcmp(inspect_sctx_user.str, user))))
          return;
      }

So looks like:

            (user && (inspect_thd->system_thread || !inspect_sctx_user.str ||
                      strcmp(inspect_sctx_user.str, user))))

the daemon appears to not have any user associated with it which results in legacy_processlist ignoring it.

Differential Revision: D54336239

fbshipit-source-id: 4fc02050e60f0ea9a1e98e07b58899b9dfb1a3b5